% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_anova.R
\name{make_anova}
\alias{make_anova}
\title{Simulate Data for ANOVA Analysis}
\usage{
make_anova(
  n_groups = 3,
  n_samples_per_group = 100,
  n_features = 100,
  n_informative = 10,
  bias = 0,
  noise = 5,
  shuffle = TRUE,
  coef = FALSE,
  random_state = NULL,
  effective_rank = NULL,
  tail_strength = 0.5,
  group_effect = 10,
  plot = FALSE
)
}
\arguments{
\item{n_groups}{Number of groups to generate data for. Default is 3.}

\item{n_samples_per_group}{Number of samples per group. Default is 100.}

\item{n_features}{Total number of features. Default is 100.}

\item{n_informative}{Number of informative features. Influences the regression target. Default is 10.}

\item{bias}{Bias term in the underlying linear model. Default is 0.0.}

\item{noise}{Standard deviation of the Gaussian noise applied to the output. Default is 5.0.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{coef}{If TRUE, the coefficients of the underlying linear model are returned. Default is FALSE.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be NULL. Default is NULL.}

\item{effective_rank}{The approximate number of singular vectors required to explain most of the input data by linear combinations. Relevant for 'make_regression'. Default is NULL.}

\item{tail_strength}{The relative importance of the fat noisy tail of the singular values profile if 'effective_rank' is not NULL. Default is 0.5.}

\item{group_effect}{The effect size to be added to each group to simulate between-group differences. Default is 10.}

\item{plot}{Logical, indicating whether to plot the dataset. Default is FALSE.}
}
\value{
A list containing three elements: 'X' is a matrix of feature data, 'Y' is the adjusted response vector including group effects, and 'GroupLabels' is a vector indicating the group of each sample.
}
\description{
This function generates synthetic data suitable for ANOVA (Analysis of Variance) testing.
It creates a dataset with a specified number of groups, samples per group, and features,
adding an optional group effect to simulate between-group differences. The function leverages
\code{make_regression} to generate the feature data (\code{X}) and then adjusts the response (\code{Y}) to include group effects.
}
\examples{
# Generate default ANOVA data
anova_data <- make_anova()

# Generate ANOVA data with 5 groups, each with 50 samples
custom_anova_data <- make_anova(n_groups = 5, n_samples_per_group = 50)
}
