% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blobs.R
\name{make_blobs}
\alias{make_blobs}
\title{Generate Blob Dataset}
\usage{
make_blobs(
  n_samples = 100,
  n_features = 2,
  centers = NULL,
  cluster_std = 1,
  center_box = c(-10, 10),
  shuffle = TRUE,
  random_state = NULL,
  return_centers = FALSE,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of samples to generate. Default is 100.}

\item{n_features}{Number of features for each sample. Default is 2.}

\item{centers}{Number of centers to generate, or the specific coordinates of the centers. If \code{NULL}, 3 centers are generated by default. Can be an integer or a matrix of center coordinates.}

\item{cluster_std}{Standard deviation of the clusters. Default is 1.0.}

\item{center_box}{A vector of length 2 indicating the bounds for the center coordinates. Default is c(-10, 10).}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be \code{NULL}. Default is \code{NULL}.}

\item{return_centers}{Logical, indicating whether to return the centers of the blobs. Default is FALSE.}

\item{plot}{Logical, indicating whether to plot the dataset. Plotting is supported for 2D and 3D data only. Default is FALSE.}
}
\value{
A list containing at least two elements: 'X' is a matrix with the coordinates of each sample, 'y' is a vector indicating the blob (center) each sample belongs to, and 'centers' are the coordinates of the blob centers if \code{return_centers} is TRUE.
}
\description{
This function generates a "blobs" dataset, useful for testing clustering algorithms. The blobs are isotropic Gaussian blobs for clustering, and the user can specify the number of centers (blobs), the standard deviation of the blobs, and the number of features for each sample. The function can return the blob centers if desired and supports plotting for 2D and 3D data.
}
\examples{
# Generate and plot a default blobs dataset
blobs_data <- make_blobs(plot = TRUE)

# Generate blobs with specific centers and plot
centers <- matrix(c(0,0, 5,5, 10,10), ncol=2, byrow=TRUE)
custom_blobs <- make_blobs(n_samples=300, centers=centers, plot=TRUE)
}
