% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_circles.R
\name{make_circles}
\alias{make_circles}
\title{Generate Circles Dataset}
\usage{
make_circles(
  n_samples = 100,
  shuffle = TRUE,
  noise = NULL,
  random_state = NULL,
  factor = 0.8,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of points to generate. Default is 100.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{noise}{Standard deviation of Gaussian noise added to the samples. If \code{NULL}, no noise is added. Default is \code{NULL}.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be \code{NULL}. Default is \code{NULL}.}

\item{factor}{Scale factor between the two circles. A smaller factor makes the inner circle smaller. Default is 0.8.}

\item{plot}{Logical, indicating whether to plot the dataset. Default is FALSE.}
}
\value{
A list with two elements: 'data' is a matrix with the coordinates of each sample, and 'labels' is a vector indicating the class (1 for outer circle, 2 for inner circle) of each sample.
}
\description{
This function generates a two-dimensional dataset comprising two circles, one within the other, known as the "circles" dataset.
It is useful for visualizing and testing algorithms that require a non-linear decision boundary, such as clustering and classification algorithms.
}
\examples{
# Generate and plot a default circles dataset
circles_data <- make_circles(plot = TRUE)

# Generate circles with noise and plot
noisy_circles <- make_circles(noise = 0.1, plot = TRUE)

# Generate circles with a different factor and plot
custom_circles <- make_circles(factor = 0.5, plot = TRUE)
}
