% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_4square.R
\name{make_four_square_dataset}
\alias{make_four_square_dataset}
\title{Generate Four Square Dataset}
\usage{
make_four_square_dataset(
  n_samples = 100,
  noise = 0,
  shuffle = TRUE,
  random_state = NULL,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of samples to generate. Default is 100.}

\item{noise}{Standard deviation of Gaussian noise added to the data. Default is 0.0.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be NULL. Default is NULL.}

\item{plot}{Logical, indicating whether to plot the dataset. Default is FALSE.}
}
\value{
A list containing two elements: 'features' is a matrix of coordinates (x, y) of the points, and 'target' is a vector indicating the class (1 or 2) of each point.
}
\description{
This function generates a synthetic dataset that consists of points distributed in four squares.
This dataset can be used for testing clustering algorithms, visualization, and other data science tasks.
The dataset is divided into two classes, with each class occupying two opposite squares.
}
\examples{
# Generate a default four square dataset and plot it
dataset <- make_four_square_dataset(n_samples=1000, plot=TRUE)

# Generate a dataset with noise and plot it
noisy_dataset <- make_four_square_dataset(n_samples=1000, noise=0.1, plot=TRUE)
}
