% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_moons.R
\name{make_moons}
\alias{make_moons}
\title{Generate Moons Dataset}
\usage{
make_moons(
  n_samples = 100,
  shuffle = TRUE,
  noise = NULL,
  random_state = NULL,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of points to generate. Default is 100.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{noise}{Standard deviation of Gaussian noise added to the samples. If \code{NULL}, no noise is added. Default is \code{NULL}.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be \code{NULL}. Default is \code{NULL}.}

\item{plot}{Logical, indicating whether to plot the dataset. Default is FALSE.}
}
\value{
A list with two elements: 'X' is a matrix with the coordinates of each sample, and 'y' is a vector indicating the class (1 or 2) of each sample.
}
\description{
This function generates a two-dimensional dataset comprising two interleaving half circles, known as the "moons" dataset. It is particularly useful for visualizing and testing clustering and classification algorithms that require a non-linear decision boundary.
}
\examples{
# Generate and plot a default moons dataset
moons_data <- make_moons(plot = TRUE)

# Generate and plot moons dataset with added noise
noisy_moons <- make_moons(noise = 0.1, plot = TRUE)

# Generate, shuffle, and plot moons dataset with specific random state
shuffled_moons <- make_moons(random_state = 42, plot = TRUE)
}
