% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regression.R
\name{make_regression}
\alias{make_regression}
\title{Generate Regression Dataset}
\usage{
make_regression(
  n_samples = 100,
  n_features = 100,
  n_informative = 10,
  n_targets = 1,
  bias = 0,
  noise = 0,
  shuffle = TRUE,
  coef = FALSE,
  random_state = NULL,
  effective_rank = NULL,
  tail_strength = 0.5,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of samples to generate. Default is 100.}

\item{n_features}{Total number of features. Non-informative features are generated as noise. Default is 100.}

\item{n_informative}{Number of informative features, i.e., features actually used to build the linear model. Default is 10.}

\item{n_targets}{Number of targets (response variables) to generate. Default is 1.}

\item{bias}{Intercept term in the underlying linear model. Default is 0.0.}

\item{noise}{Standard deviation of Gaussian noise added to the output. Default is 0.0.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{coef}{Logical, indicates whether to return the coefficients of the underlying linear model along with the dataset. Default is FALSE.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be \code{NULL}. Default is \code{NULL}.}

\item{effective_rank}{The approximate number of singular vectors required to explain most of the input data by linear combinations. Relevant for simulating data with collinear features. Default is \code{NULL}.}

\item{tail_strength}{The relative importance of the fat noisy tail of the singular values profile if \code{effective_rank} is not \code{NULL}. Affects how quickly the importance of singular vectors falls off. Default is 0.5.}

\item{plot}{Logical, indicating whether to plot the dataset. Supported only for datasets with a single target. Default is FALSE.}
}
\value{
A list containing the feature matrix 'X', the target vector 'Y', and optionally the true coefficients if \code{coef} is TRUE.
}
\description{
This function generates a regression dataset suitable for testing linear regression models. It allows for customization of the number of samples, features, informative features, targets, noise, and other parameters to simulate various regression scenarios.
}
\examples{
# Generate a simple regression dataset and plot it
simple_regression <- make_regression(plot = TRUE)

# Generate a regression dataset with more informative features and plot it
informative_regression <- make_regression(n_features=200, n_informative=50, plot=TRUE)

# Generate a multi-output regression dataset and plot it
multi_output_regression <- make_regression(n_targets=2, plot=TRUE)
}
