% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spiral.R
\name{make_spiral}
\alias{make_spiral}
\title{Generate Spiral Dataset}
\usage{
make_spiral(
  n_samples = 100,
  shuffle = TRUE,
  noise = NULL,
  random_state = NULL,
  plot = FALSE
)
}
\arguments{
\item{n_samples}{Total number of points to generate. Default is 100.}

\item{shuffle}{Logical, indicating whether to shuffle the samples. Default is TRUE.}

\item{noise}{Standard deviation of Gaussian noise added to the samples. If \code{NULL}, no noise is added. Default is \code{NULL}.}

\item{random_state}{Integer, the seed used by the random number generator for reproducibility. Can be \code{NULL}. Default is \code{NULL}.}

\item{plot}{Logical, indicating whether to plot the dataset. Default is FALSE.}
}
\value{
A list with two elements: 'X' is a matrix with the coordinates of each sample, and 'y' is a vector indicating the class (1 or 2) of each sample.
}
\description{
This function generates a two-dimensional dataset comprising two interwoven spirals. It's useful for visualizing and testing algorithms that require learning complex, non-linear decision boundaries, such as neural networks or advanced clustering techniques.
}
\examples{
# Generate and plot a default spiral dataset
spiral_data <- make_spiral(plot = TRUE)

# Generate and plot a spiral dataset with added noise
noisy_spiral <- make_spiral(n_samples = 200, noise = 0.1, plot = TRUE)

# Generate a spiral dataset with a specific random state and plot it
spiral_seed <- make_spiral(n_samples = 200, noise = 0.1, random_state = 123, plot = TRUE)
}
